+++ projects/gloo/pkg/syncer/perf_test/perf_suite_test.go
@@ -3,9 +3,10 @@ package perf_test
 import (
 	"testing"
 
+	"github.com/solo-io/gloo/test/ginkgo/labels"
+
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/solo-io/gloo/test/gomega/labels"
 )
 
 func TestPerformance(t *testing.T) {
