+++ test/e2e/example_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/solo-io/gloo/test/gomega/labels"
+	"github.com/solo-io/gloo/test/ginkgo/labels"
 
 	"github.com/solo-io/gloo/test/testutils"
